fix(test): #6060 source-side 단위 시험을 빼고 SVG 쪽 번호를 u32 로 맞춘다
CI Lint 가 src/renderer/layout/text_measurement.rs::tests 에 시험 2건을 더해 4223>4221 로 거부했다.
전각 낫표 계약은 tests/cases/issue_6060_fullwidth_cjk_quote.rs SVG 간격 시험이 가드한다.
PAGE 를 u32 로 바꿔 archive-b·prop roundtrip 의 E0308 을 고친다.

+++ tests/cases/issue_6060_fullwidth_cjk_quote.rs
@@ -14,7 +14,7 @@ use std::path::Path;
 use rhwp::wasm_api::HwpDocument;
 
 const SAMPLE: &str = "samples/hwp3-sample16-hwp5.hwpx";
-const PAGE: usize = 4; // 5쪽
+const PAGE: u32 = 4; // 5쪽
 const NEEDLE: &str = "「국가를당사자";
 
 fn load_doc() -> HwpDocument {
